Not tracking fax_file by default
As `fuzzy_for_options` not working at the raw_record level plugins. Tracking fax_file makes it harder to load multiple runs of simulations with the same context. Another option is to move the read instruction into the `compute` method so that run_id can be used to complete the `fax_file`. But that is too much work compared to this quick fix.